refactor and add club attributes
